crawler/requirements.txt
selenium==4.17.2
beautifulsoup4==4.12.3
requests==2.31.0
youtube-transcript-api==0.6.2
qdrant-client==1.7.0
sentence-transformers==2.3.1
python-dotenv==1.0.0
pandas==2.2.0
tqdm==4.66.1
webdriver-manager==4.0.1
